Move file for DatasetPropertyNew component
